+++ src/account.c
@@ -19,7 +19,7 @@ static atomic_int  save_id;
 
 static ALLEGRO_THREAD *account_thread = NULL;
 
-int account_money = 0;
+float account_money = 0;
 
 static const long ONE_HOUR = 3600;
